Merge to Main Branch

+++ controllers/BookmarkController.js
@@ -11,10 +11,56 @@ class BookmarkController {
     const result = await prisma.bookmarks.findMany({
       where: { authorId },
       include: {
-        recipe: true,
+        recipe: {
+          include: {
+            author: {
+              select: {
+                name: true,
+                photo: true,
+                slug: true,
+              },
+            },
+            secondCategory: {
+              select: {
+                id: true,
+                name: true,
+                photo: true,
+              },
+            },
+            utensils: {
+              select: {
+                utensil: {
+                  select: {
+                    id: true,
+                    name: true,
+                    photo: true,
+                  },
+                },
+              },
+            },
+            _count: {
+              select: { bookmarks: true },
+            },
+          },
+        },
       },
     });
-    res.status(200).json(result);
+
+    const bookmark = result.map((data) => {
+      data.recipe.totalBookmark = data.recipe._count.bookmarks;
+
+      const utensils = data.recipe.utensils.map((item) => {
+        return item.utensil;
+      });
+      data.recipe.utensils = utensils;
+
+      const { _count, searchIngredients, ...restRecipe } = data.recipe;
+      data.recipe = restRecipe;
+
+      return data;
+    });
+
+    res.status(200).json(bookmark);
   }
 
   static async store(req, res) {
@@ -43,13 +89,24 @@ class BookmarkController {
   }
 
   static async destroy(req, res, next) {
+    const recipeId = req.params.recipeId;
+    const authorId = req.loggedUser.id;
+
     try {
-      await prisma.bookmarks.delete({
+      const bookmark = await prisma.bookmarks.findFirst({
         where: {
-          id: req.params.id,
+          AND: [{ recipeId }, { authorId }],
         },
       });
-      res.status(200).json({ message: 'Bookmark deleted' });
+
+      if (bookmark) {
+        await prisma.bookmarks.delete({
+          where: { id: bookmark.id },
+        });
+        res.status(200).json({ message: 'Bookmark deleted' });
+      } else {
+        res.status(400).json({ message: 'Data not found' });
+      }
     } catch (error) {
       next(error);
     }
